wip: refactor starting player logic
I'm trying to use the 'deck' function `getCard` to get the 2 of clubs
and grab the player id from that card's `'location_arg'`. But I'm getting
errors in BGA, saying I'm trying to active a player that's not in the game.
Perhaps it's not getting the right card, or perhaps the player is not
being returned in the format I expect. Leaving it for another night because
I'm sleepy.